requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==2.3.3
Flask-MySQLdb==2.0.0
fsspec==2023.9.2
huggingface-hub==0.16.4
idna==3.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.3.3
sympy==1.12
tokenizers==0.14.0
torch==2.0.1
urllib3==2.0.6
Werkzeug==2.3.7
